rename hook/patch class name and file location
