+++ lib/redmine/acts/heartable.rb
